DBot/requirements.txt
discord
python-dotenv
pyyaml
requests
pynacl
pytube
moviepy